Created database, user handling, html pages, notes
The database schema is created, where users can write notes on the website.
The basic HTML pages are created for home, login and sign-up.
The user can delete notes.
Authorization mechanics are added where a user can create an account and login and can't access certain pages when logged in or logged out.